renamed folders in reusable and module components
