+++ app/blog/page.tsx
@@ -1,4 +1,4 @@
-import Tab from '../../components/reusable-components/Tab/tab';
+import Tab from '../../components/reusable-components/tab/Tab';
 
 const Blog = () => {
   return <Tab tabOneText="За тебе" tabTwoText="Најново" />;
